Kext: mechanically re-port the remaining older hand-written methods (batches 2-4, 112 fns) from the stock decompile
Also fixes found while porting: unsigned-int* out-params of Shared::new_texture/new_global_texture were
aliased as byte pointers (stores would have been 1-byte writes); Tools/audit_cast_types.py now guards that.
Ledger 506/506, vtable diff clean, callee compare at -O0 shows only compiler-artifact differences.

+++ Sources/IOATIR500Shared_new_global_texture_Port.cpp
@@ -30,7 +30,7 @@ extern "C" UInt32 GH_OSIncrementAtomic(...) asm("_OSIncrementAtomic");
 /* real addr 0x17740 */
 VendorTextureBuffer *IOATIR500Shared::new_global_texture(UInt32 param_1, unsigned int *real_param_2) {
     UInt8 *self = reinterpret_cast<UInt8 *>(this);
-    UInt8 *param_2 = reinterpret_cast<UInt8 *>(real_param_2);
+    unsigned int *param_2 = reinterpret_cast<unsigned int *>(real_param_2);
 
   UInt8 *pIVar1;
   int *piVar2;
